pyproject.toml
[project]
name = "stuart-bot-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "loguru>=0.7.3",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "archpython-cli",
    "black>=25.9.0",
    "pyright>=1.1.405",
    "ruff>=0.13.1",
    "taskipy>=1.14.1",
]
[tool.taskipy.tasks]
lint = "ruff check src --fix && black src && isort src"
typecheck = "pyright src"
run_dev = "uv run src/main.py"

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["stuart-api"]
combine_as_imports = true

[tool.uv.sources]
archpython-cli = { git = "https://github.com/Filipi1/archpython-cli" }
